revert for rename `asserts.segment.expression`.
